requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.7
tabulate>=0.9.0
colorama>=0.4.6
rich>=13.7.0
watchdog>=3.0.0
python-daemon>=3.0.1
lockfile>=0.12.2
pygithub>=2.1.1
requests>=2.31.0